[trusted-firmware-m]: Updated to e7efdc6
