+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/ipc/include/tfm_message_queue.h
@@ -7,9 +7,7 @@
 #ifndef __TFM_MESSAGE_QUEUE_H__
 #define __TFM_MESSAGE_QUEUE_H__
 
-#ifndef TFM_MSG_QUEUE_MAX_MSG_NUM
-#define TFM_MSG_QUEUE_MAX_MSG_NUM   128
-#endif
+#define TFM_MSG_QUEUE_MAX_MSG_NUM   16
 #define TFM_MSG_MAGIC               0x15154343
 /* Message struct to collect parameter from client */
 struct tfm_msg_body_t {
