+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/ipc/include/tfm_spm_signal_defs.h
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2018-2019, Arm Limited. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ */
+#ifndef __TFM_SPM_SIGNAL_DEFS_H__
+#define __TFM_SPM_SIGNAL_DEFS_H__
+
+#include "test/test_services/tfm_ipc_service/tfm_ipc_service_partition.h"
+#include "test/test_services/tfm_core_test/tfm_ss_core_test_signal.h"
+#include "test/test_services/tfm_core_test_2/tfm_ss_core_test_2_signal.h"
+#include "secure_fw/services/crypto/tfm_crypto_signal.h"
+#include "test/test_services/tfm_secure_client_service/tfm_sec_client_ser_sig.h"
+#include "secure_fw/services/secure_storage/tfm_sst_signal.h"
+#include "secure_fw/services/initial_attestation/tfm_attest_signal.h"
+
+#endif /* __TFM_SPM_SIGNAL_DEFS_H__ */
