+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/ipc/psa_client.c
@@ -10,7 +10,7 @@
 #include "tfm_svc.h"
 #include "psa_client.h"
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 uint32_t psa_framework_version(void)
 {
     __ASM volatile("SVC %0           \n"
@@ -18,7 +18,7 @@ uint32_t psa_framework_version(void)
                    : : "I" (TFM_SVC_PSA_FRAMEWORK_VERSION));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 uint32_t psa_version(uint32_t sid)
 {
     __ASM volatile("SVC %0           \n"
@@ -26,7 +26,7 @@ uint32_t psa_version(uint32_t sid)
                    : : "I" (TFM_SVC_PSA_VERSION));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 psa_handle_t psa_connect(uint32_t sid, uint32_t minor_version)
 {
     __ASM volatile("SVC %0           \n"
@@ -34,7 +34,7 @@ psa_handle_t psa_connect(uint32_t sid, uint32_t minor_version)
                    : : "I" (TFM_SVC_PSA_CONNECT));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 psa_status_t psa_call(psa_handle_t handle,
                       const psa_invec *in_vec,
                       size_t in_len,
@@ -46,7 +46,7 @@ psa_status_t psa_call(psa_handle_t handle,
                    : : "I" (TFM_SVC_PSA_CALL));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_close(psa_handle_t handle)
 {
     __ASM volatile("SVC %0           \n"
