+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/ipc/psa_service.c
@@ -11,7 +11,7 @@
 #include "psa_client.h"
 #include "psa_service.h"
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 psa_signal_t psa_wait(psa_signal_t signal_mask, uint32_t timeout)
 
 {
@@ -20,7 +20,7 @@ psa_signal_t psa_wait(psa_signal_t signal_mask, uint32_t timeout)
                    : : "I" (TFM_SVC_PSA_WAIT));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 psa_status_t psa_get(psa_signal_t signal, psa_msg_t *msg)
 {
     __ASM volatile("SVC %0           \n"
@@ -28,7 +28,7 @@ psa_status_t psa_get(psa_signal_t signal, psa_msg_t *msg)
                    : : "I" (TFM_SVC_PSA_GET));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_set_rhandle(psa_handle_t msg_handle, void *rhandle)
 {
     __ASM volatile("SVC %0           \n"
@@ -36,7 +36,7 @@ void psa_set_rhandle(psa_handle_t msg_handle, void *rhandle)
                    : : "I" (TFM_SVC_PSA_SET_RHANDLE));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 size_t psa_read(psa_handle_t msg_handle, uint32_t invec_idx,
                 void *buffer, size_t num_bytes)
 
@@ -46,7 +46,7 @@ size_t psa_read(psa_handle_t msg_handle, uint32_t invec_idx,
                    : : "I" (TFM_SVC_PSA_READ));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 size_t psa_skip(psa_handle_t msg_handle, uint32_t invec_idx, size_t num_bytes)
 {
     __ASM volatile("SVC %0           \n"
@@ -54,7 +54,7 @@ size_t psa_skip(psa_handle_t msg_handle, uint32_t invec_idx, size_t num_bytes)
                    : : "I" (TFM_SVC_PSA_SKIP));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_write(psa_handle_t msg_handle, uint32_t outvec_idx,
                const void *buffer, size_t num_bytes)
 {
@@ -63,7 +63,7 @@ void psa_write(psa_handle_t msg_handle, uint32_t outvec_idx,
                    : : "I" (TFM_SVC_PSA_WRITE));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_reply(psa_handle_t msg_handle, psa_status_t retval)
 {
     __ASM volatile("SVC %0           \n"
@@ -71,7 +71,7 @@ void psa_reply(psa_handle_t msg_handle, psa_status_t retval)
                    : : "I" (TFM_SVC_PSA_REPLY));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_notify(int32_t partition_id)
 {
     __ASM volatile("SVC %0           \n"
@@ -79,7 +79,7 @@ void psa_notify(int32_t partition_id)
                    : : "I" (TFM_SVC_PSA_NOTIFY));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_clear(void)
 {
     __ASM volatile("SVC %0           \n"
@@ -87,7 +87,7 @@ void psa_clear(void)
                    : : "I" (TFM_SVC_PSA_CLEAR));
 }
 
-__attribute__((naked))
+__attribute__((naked, section("SFN")))
 void psa_eoi(psa_signal_t irq_signal)
 {
     __ASM volatile("SVC %0           \n"
