+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/ipc/tfm_svcalls.c
@@ -21,6 +21,7 @@
 #include "tfm_api.h"
 #include "tfm_secure_api.h"
 #include "tfm_memory_utils.h"
+#include "spm_api.h"
 
 #define PSA_TIMEOUT_MASK        PSA_BLOCK
 
@@ -116,9 +117,18 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
     struct tfm_spm_service_t *service;
     struct tfm_msg_body_t *msg;
     int i;
+    struct tfm_spm_ipc_partition_t *partition = NULL;
+    uint32_t privileged;
 
     TFM_ASSERT(args != NULL);
     handle = (psa_handle_t)args[0];
+
+    partition = tfm_spm_get_running_partition();
+    if (!partition) {
+        tfm_panic();
+    }
+    privileged = tfm_spm_partition_get_privileged_mode(partition->index);
+
     if (!ns_caller) {
         inptr = (psa_invec *)args[1];
         in_num = (size_t)args[2];
@@ -153,11 +163,11 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
          * memory reference for buffer is invalid or not readable.
          */
         if (tfm_memory_check((void *)args[1], sizeof(uint32_t),
-            ns_caller, TFM_MEMORY_ACCESS_RO) != IPC_SUCCESS) {
+            ns_caller, TFM_MEMORY_ACCESS_RO, privileged) != IPC_SUCCESS) {
             tfm_panic();
         }
         if (tfm_memory_check((void *)args[2], sizeof(uint32_t),
-            ns_caller, TFM_MEMORY_ACCESS_RO) != IPC_SUCCESS) {
+            ns_caller, TFM_MEMORY_ACCESS_RO, privileged) != IPC_SUCCESS) {
             tfm_panic();
         }
 
@@ -185,7 +195,7 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
      * readable.
      */
     if (tfm_memory_check((void *)inptr, in_num * sizeof(psa_invec),
-        ns_caller, TFM_MEMORY_ACCESS_RO) != IPC_SUCCESS) {
+        ns_caller, TFM_MEMORY_ACCESS_RO, privileged) != IPC_SUCCESS) {
         tfm_panic();
     }
     /*
@@ -194,7 +204,7 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
      * the wrap output vector is invalid or not read-write.
      */
     if (tfm_memory_check((void *)outptr, out_num * sizeof(psa_outvec),
-        ns_caller, TFM_MEMORY_ACCESS_RW) != IPC_SUCCESS) {
+        ns_caller, TFM_MEMORY_ACCESS_RW, privileged) != IPC_SUCCESS) {
         tfm_panic();
     }
 
@@ -211,7 +221,7 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
      */
     for (i = 0; i < in_num; i++) {
         if (tfm_memory_check((void *)invecs[i].base, invecs[i].len,
-            ns_caller, TFM_MEMORY_ACCESS_RO) != IPC_SUCCESS) {
+            ns_caller, TFM_MEMORY_ACCESS_RO, privileged) != IPC_SUCCESS) {
             tfm_panic();
         }
     }
@@ -221,7 +231,7 @@ psa_status_t tfm_svcall_psa_call(uint32_t *args, int32_t ns_caller, uint32_t lr)
      */
     for (i = 0; i < out_num; i++) {
         if (tfm_memory_check(outvecs[i].base, outvecs[i].len,
-            ns_caller, TFM_MEMORY_ACCESS_RW) != IPC_SUCCESS) {
+            ns_caller, TFM_MEMORY_ACCESS_RW, privileged) != IPC_SUCCESS) {
             tfm_panic();
         }
     }
@@ -363,6 +373,7 @@ static psa_status_t tfm_svcall_psa_get(uint32_t *args)
     struct tfm_spm_service_t *service = NULL;
     struct tfm_msg_body_t *tmp_msg = NULL;
     struct tfm_spm_ipc_partition_t *partition = NULL;
+    uint32_t privileged;
 
     TFM_ASSERT(args != NULL);
     signal = (psa_signal_t)args[0];
@@ -376,17 +387,18 @@ static psa_status_t tfm_svcall_psa_get(uint32_t *args)
         tfm_panic();
     }
 
+    partition = tfm_spm_get_running_partition();
+    if (!partition) {
+        tfm_panic();
+    }
+    privileged = tfm_spm_partition_get_privileged_mode(partition->index);
+
     /*
      * Write the message to the service buffer. It is a fatal error if the
      * input msg pointer is not a valid memory reference or not read-write.
      */
     if (tfm_memory_check((void *)msg, sizeof(psa_msg_t),
-        false, TFM_MEMORY_ACCESS_RW) != IPC_SUCCESS) {
-        tfm_panic();
-    }
-
-    partition = tfm_spm_get_running_partition();
-    if (!partition) {
+        false, TFM_MEMORY_ACCESS_RW, privileged) != IPC_SUCCESS) {
         tfm_panic();
     }
 
@@ -499,6 +511,8 @@ static size_t tfm_svcall_psa_read(uint32_t *args)
     size_t num_bytes;
     size_t bytes;
     struct tfm_msg_body_t *msg = NULL;
+    uint32_t privileged;
+    struct tfm_spm_ipc_partition_t *partition = NULL;
 
     TFM_ASSERT(args != NULL);
     msg_handle = (psa_handle_t)args[0];
@@ -511,6 +525,9 @@ static size_t tfm_svcall_psa_read(uint32_t *args)
     if (!msg) {
         tfm_panic();
     }
+
+    partition = msg->service->partition;
+    privileged = tfm_spm_partition_get_privileged_mode(partition->index);
 
     /*
      * It is a fatal error if message handle does not refer to a PSA_IPC_CALL
@@ -538,7 +555,7 @@ static size_t tfm_svcall_psa_read(uint32_t *args)
      * if the memory reference for buffer is invalid or not read-write.
      */
     if (tfm_memory_check(buffer, num_bytes, false,
-        TFM_MEMORY_ACCESS_RW) != IPC_SUCCESS) {
+        TFM_MEMORY_ACCESS_RW, privileged) != IPC_SUCCESS) {
         tfm_panic();
     }
 
@@ -650,6 +667,8 @@ static void tfm_svcall_psa_write(uint32_t *args)
     void *buffer = NULL;
     size_t num_bytes;
     struct tfm_msg_body_t *msg = NULL;
+    uint32_t privileged;
+    struct tfm_spm_ipc_partition_t *partition = NULL;
 
     TFM_ASSERT(args != NULL);
     msg_handle = (psa_handle_t)args[0];
@@ -662,6 +681,9 @@ static void tfm_svcall_psa_write(uint32_t *args)
     if (!msg) {
         tfm_panic();
     }
+
+    partition = msg->service->partition;
+    privileged = tfm_spm_partition_get_privileged_mode(partition->index);
 
     /*
      * It is a fatal error if message handle does not refer to a PSA_IPC_CALL
@@ -693,7 +715,7 @@ static void tfm_svcall_psa_write(uint32_t *args)
      * if the memory reference for buffer is invalid or not readable.
      */
     if (tfm_memory_check(buffer, num_bytes, false,
-        TFM_MEMORY_ACCESS_RO) != IPC_SUCCESS) {
+        TFM_MEMORY_ACCESS_RO, privileged) != IPC_SUCCESS) {
         tfm_panic();
     }
 
@@ -956,6 +978,9 @@ int32_t SVC_Handler_IPC(tfm_svc_number_t svc_num, uint32_t *ctx, uint32_t lr)
     switch (svc_num) {
     case TFM_SVC_SCHEDULE:
         tfm_thrd_activate_schedule();
+        break;
+    case TFM_SVC_EXIT_THRD:
+        tfm_svcall_thrd_exit();
         break;
     case TFM_SVC_PSA_FRAMEWORK_VERSION:
         return tfm_svcall_psa_framework_version();
