+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/tfm_secure_api.h
@@ -96,14 +96,19 @@ int32_t tfm_core_sfn_request_thread_mode(struct tfm_sfn_req_s *desc_ptr);
  * This function assumes, that the current MPU configuration is set for the
  * partition to be checked.
  *
- * \param[in] p          The start address of the range to check
- * \param[in] s          The size of the range to check
- * \param[in] ns_caller  Whether the current partition is a non-secure one
+ * \param[in] p                The start address of the range to check
+ * \param[in] s                The size of the range to check
+ * \param[in] ns_caller        Whether the current partition is a non-secure one
+ * \param[in] privileged       Privileged mode or unprivileged mode:
+ *                             \ref TFM_PARTITION_UNPRIVILEGED_MODE
+ *                             \ref TFM_PARTITION_PRIVILEGED_MODE
  *
- * \return 1 if the partition has access to the memory range, 0 otherwise.
+ * \return TFM_SUCCESS if the partition has access to the memory range,
+ *         TFM_ERROR_GENERIC otherwise.
  */
 int32_t tfm_core_has_read_access_to_region(const void *p, size_t s,
-                                           uint32_t ns_caller);
+                                           uint32_t ns_caller,
+                                           uint32_t privileged);
 
 /**
  * \brief Check whether the current partition has write access to a memory range
@@ -111,15 +116,39 @@ int32_t tfm_core_has_read_access_to_region(const void *p, size_t s,
  * This function assumes, that the current MPU configuration is set for the
  * partition to be checked.
  *
- * \param[in] p          The start address of the range to check
- * \param[in] s          The size of the range to check
- * \param[in] ns_caller  Whether the current partition is a non-secure one
+ * \param[in] p                The start address of the range to check
+ * \param[in] s                The size of the range to check
+ * \param[in] ns_caller        Whether the current partition is a non-secure one
+ * \param[in] privileged       Privileged mode or unprivileged mode:
+ *                             \ref TFM_PARTITION_UNPRIVILEGED_MODE
+ *                             \ref TFM_PARTITION_PRIVILEGED_MODE
  *
- * \return 1 if the partition has access to the memory range, 0 otherwise.
+ * \return TFM_SUCCESS if the partition has access to the memory range,
+ *         TFM_ERROR_GENERIC otherwise.
  */
 int32_t tfm_core_has_write_access_to_region(void *p, size_t s,
-                                            uint32_t ns_caller);
+                                            uint32_t ns_caller,
+                                            uint32_t privileged);
 
+#ifdef TFM_PSA_API
+/* The following macros are only valid if secure services can be called
+ * using veneer functions. This is not the case if IPC messaging is enabled
+ */
+#define TFM_CORE_IOVEC_SFN_REQUEST(id, fn, a, b, c, d)               \
+        do {                                                         \
+            ERROR_MSG("Invalid TF-M configuration detected");        \
+            tfm_secure_api_error_handler();                          \
+            /* This point never reached */                           \
+            return (int32_t)TFM_ERROR_GENERIC;                       \
+        } while (0)
+#define TFM_CORE_SFN_REQUEST(id, fn, a, b, c, d)                     \
+        do {                                                         \
+            ERROR_MSG("Invalid TF-M configuration detected");        \
+            tfm_secure_api_error_handler();                          \
+            /* This point never reached */                           \
+            return (int32_t)TFM_ERROR_GENERIC;                       \
+        } while (0)
+#else
 #define TFM_CORE_IOVEC_SFN_REQUEST(id, fn, a, b, c, d) \
         return tfm_core_partition_request(id, fn, TFM_SFN_API_IOVEC, \
                 (int32_t)a, (int32_t)b, (int32_t)c, (int32_t)d)
@@ -136,7 +165,7 @@ int32_t tfm_core_partition_request(uint32_t id, void *fn, int32_t iovec_api,
     struct tfm_sfn_req_s desc, *desc_ptr = &desc;
 
     desc.sp_id = id;
-    desc.sfn = fn;
+    desc.sfn = (sfn_t) fn;
     desc.args = args;
     /*
      * This preprocessor condition checks if a version of GCC smaller than
@@ -180,5 +209,6 @@ int32_t tfm_core_partition_request(uint32_t id, void *fn, int32_t iovec_api,
 
     }
 }
+#endif
 
 #endif /* __TFM_SECURE_API_H__ */
