+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/core/tfm_svc.h
@@ -23,6 +23,7 @@ typedef enum {
 #ifdef TFM_PSA_API
     TFM_SVC_IPC_REQUEST,
     TFM_SVC_SCHEDULE,
+    TFM_SVC_EXIT_THRD,
     /* PSA Client SVC */
     TFM_SVC_PSA_FRAMEWORK_VERSION,
     TFM_SVC_PSA_VERSION,
