+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/spm/spm_db.h
@@ -21,8 +21,6 @@ typedef psa_status_t(*sp_init_function)(void);
 #define TFM_PARTITION_TYPE_APP   "APPLICATION-ROT"
 #define TFM_PARTITION_TYPE_PSA   "PSA-ROT"
 
-#define TFM_STACK_SIZE           (1024 * 5)
-
 #ifdef TFM_PSA_API
 enum tfm_partition_priority {
     TFM_PRIORITY_LOW = THRD_PRIOR_LOWEST,
@@ -59,23 +57,18 @@ struct spm_partition_desc_t {
     struct spm_partition_static_data_t static_data;
     struct spm_partition_runtime_data_t runtime_data;
     struct tfm_spm_partition_platform_data_t *platform_data;
-#if TFM_LVL != 1
+#if (TFM_LVL != 1) || defined(TFM_PSA_API)
     struct tfm_spm_partition_memory_data_t memory_data;
 #endif
 #ifdef TFM_PSA_API
     struct tfm_thrd_ctx sp_thrd;
-    /*
-     * stack_limit points to starting address of the partitions' stack plus the partitions' stack size.
-     */
-    uint32_t stack_limit;
-    uint32_t stack_size;
 #endif
 };
 
 /* Macros to pick linker symbols and allow to form the partition data base */
 #define REGION(a, b, c) a##b##c
 #define REGION_NAME(a, b, c) REGION(a, b, c)
-#if TFM_LVL == 1
+#if (TFM_LVL == 1) && !defined(TFM_PSA_API)
 #define REGION_DECLARE(a, b, c)
 #else
 #define REGION_DECLARE(a, b, c) extern uint32_t REGION_NAME(a, b, c)
