+++ components/TARGET_PSA/TARGET_TFM/COMPONENT_SPE/secure_fw/spm/spm_partition_defs.h
@@ -29,7 +29,7 @@
  */
 #define TFM_SP_CORE_ID (1)
 
-#include "tfm_partition_defs.inc"
+#include "secure_fw/services/tfm_partition_defs.inc"
 
 /* This limit is only used to define the size of the database reserved for
  * partitions. There's no requirement that it match the number of partitions
